Math.Algebra.Field.Base

Documentation

newtype Q

Q is just the rationals, but with a better show function than the Prelude version

Constructors

Q Rational

numeratorQ :: Q -> Integer

denominatorQ :: Q -> Integer

extendedEuclid :: Integral t => t -> t -> (t, t, t)

newtype Fp n

Constructors

Fp Integer

class (Eq fq, Fractional fq) => FiniteField fq where

Minimal complete definition

eltsFq, basisFq

Methods

eltsFq :: fq -> [fq]

basisFq :: fq -> [fq]

primitiveElt :: (Eq a, Num a) => [a] -> a

powers :: (Num a, Eq a) => a -> [a]

char :: Foldable t => t a -> Int

type F2 = Fp T2

F2 is a type for the finite field with 2 elements

f2 :: [F2]

f2 lists the elements of F2

type F3 = Fp T3

F3 is a type for the finite field with 3 elements

f3 :: [F3]

f3 lists the elements of F3

type F5 = Fp T5

F5 is a type for the finite field with 5 elements

f5 :: [F5]

f5 lists the elements of F5

type F7 = Fp T7

F7 is a type for the finite field with 7 elements

f7 :: [F7]

f7 lists the elements of F7

type F11 = Fp T11

f11 :: [F11]

type F13 = Fp T13

f13 :: [F13]

type F17 = Fp T17

f17 :: [F17]

type F19 = Fp T19

f19 :: [F19]

type F23 = Fp T23

f23 :: [F23]

type F29 = Fp T29

f29 :: [F29]

type F31 = Fp T31

f31 :: [F31]

type F37 = Fp T37

f37 :: [F37]

type F41 = Fp T41

f41 :: [F41]

type F43 = Fp T43

f43 :: [F43]

type F47 = Fp T47

f47 :: [F47]

type F53 = Fp T53

f53 :: [F53]

type F59 = Fp T59

f59 :: [F59]

type F61 = Fp T61

f61 :: [F61]

type F67 = Fp T67

f67 :: [F67]

type F71 = Fp T71

f71 :: [F71]

type F73 = Fp T73

f73 :: [F73]

type F79 = Fp T79

f79 :: [F79]

type F83 = Fp T83

f83 :: [F83]

type F89 = Fp T89

f89 :: [F89]

type F97 = Fp T97

f97 :: [F97]