Math.Algebra.Group.RandomSchreierSims

Documentation

testProdRepl :: IO ()

initProdRepl :: (Ord a, Show a) => [Permutation a] -> IO (Int, IOArray Int (Permutation a))

nextProdRepl :: (Ord a, Show a) => (Int, IOArray Int (Permutation a)) -> IO (Maybe (Permutation a))

updateArray :: (Num a1, Num i, HasInverses a1, Integral t, Ix i, MArray a a1 m) => a i a1 -> i -> i -> t -> m (Maybe a1)

sgs :: (Ord a, Show a) => [Permutation a] -> [Permutation a]

Given generators for a permutation group, return a strong generating set. The result is calculated using random Schreier-Sims algorithm, so has a small (<10^-6) chance of being incomplete. The sgs is relative to the base implied by the Ord instance.

rss :: (Show k, Ord k) => [Permutation k] -> [((k, Map k (Permutation k)), [Permutation k])]

rss' :: (Eq t, Show a, Ord a, Num t) => (Int, IOArray Int (Permutation a)) -> [((a, Map a (Permutation a)), [Permutation a])] -> t -> IO [((a, Map a (Permutation a)), [Permutation a])]

initLevels :: (Foldable t1, Ord k, Num a) => t1 (Permutation k) -> [((k, Map k a), [t])]

updateLevels :: Ord c => [((c, Map c (Permutation c)), [Permutation c])] -> Maybe (Permutation c) -> (Bool, [((c, Map c (Permutation c)), [Permutation c])])

updateLevels' :: Ord t => [((t, Map t (Permutation t)), [Permutation t])] -> [((t, Map t (Permutation t)), [Permutation t])] -> Permutation t -> t -> [((t, Map t (Permutation t)), [Permutation t])]

baseTransversalsSGS :: Ord k => [Permutation k] -> [(k, Map k (Permutation k))]

isMemberSGS :: (Ord a, Show a) => [Permutation a] -> Permutation a -> Bool

Given a strong generating set gs, isMemberSGS gs is a membership test for the group