Math.Algebra.Group.SchreierSims

Documentation

cosetRepsGx :: Ord k => [Permutation k] -> k -> Map k (Permutation k)

schreierGeneratorsGx :: Ord t => (t, Map t (Permutation t)) -> [Permutation t] -> [Permutation t]

sift :: Ord k => [(k, Map k (Permutation k))] -> Permutation k -> Maybe (Permutation k)

findBase :: Ord a => [Permutation a] -> a

sgs :: (Ord a, Show a) => [Permutation a] -> [Permutation a]

Given generators for a permutation group, return a strong generating set. The result is calculated using Schreier-Sims algorithm, and is relative to the base implied by the Ord instance

bsgs :: Ord t => [Permutation t] -> [(t, Map t (Permutation t))]

bsgs' :: Ord t => [t] -> [Permutation t] -> [(t, Map t (Permutation t))]

newLevel :: Ord t => [t] -> [Permutation t] -> ([t], ((t, Map t (Permutation t)), [Permutation t]))

newLevel' :: Ord t => t -> [Permutation t] -> ((t, Map t (Permutation t)), [Permutation t])

ss :: Ord t => [t] -> [Permutation t] -> [((t, Map t (Permutation t)), [Permutation t])]

ss' :: Ord t => [t] -> [((t, Map t (Permutation t)), [Permutation t])] -> [((t, Map t (Permutation t)), [Permutation t])] -> [((t, Map t (Permutation t)), [Permutation t])]

isMemberBSGS :: Ord k => [(k, Map k (Permutation k))] -> Permutation k -> Bool

eltsBSGS :: Num b => [(a, Map k b)] -> [b]

cartProd :: [[a]] -> [[a]]

orderBSGS :: [(a1, Map k a)] -> Integer

isMember :: (Ord t, Show t) => [Permutation t] -> Permutation t -> Bool

Given generators for a group, determine whether a permutation is a member of the group, using Schreier-Sims algorithm

elts :: (Ord t, Show t) => [Permutation t] -> [Permutation t]

Given generators for a group, return a (sorted) list of all elements of the group, using Schreier-Sims algorithm

order :: (Ord t, Show t) => [Permutation t] -> Integer

Given generators for a group, return the order of the group (the number of elements), using Schreier-Sims algorithm

isSubgp :: (Ord k, Foldable t) => t (Permutation k) -> [Permutation k] -> Bool

isNormal :: Ord k => [Permutation k] -> [Permutation k] -> Bool

index :: (Show t, Show t1, Ord t, Ord t1) => [Permutation t1] -> [Permutation t] -> Integer

reduceGens :: Ord t => [Permutation t] -> [Permutation t]

reduceGensBSGS :: Ord t => [Permutation t] -> ([Permutation t], [(t, Map t (Permutation t))])

normalClosure :: Ord t => [Permutation t] -> [Permutation t] -> [Permutation t]

commutatorGp :: Ord t => [Permutation t] -> [Permutation t] -> [Permutation t]

derivedSubgp :: Ord t => [Permutation t] -> [Permutation t]