Math.Algebra.Group.StringRewriting

Documentation

rewrite :: Eq a => [([a], [a])] -> [a] -> [a]

Given a list of rewrite rules of the form (left,right), and a word, rewrite it by repeatedly replacing any left substring in the word by the corresponding right

rewrite1 :: Eq a => ([a], [a]) -> [a] -> Maybe [a]

splitSubstring :: Eq a => [a] -> [a] -> Maybe ([a], [a])

findOverlap :: Eq a => [a] -> [a] -> Maybe ([a], [a], [a])

knuthBendix1 :: Ord a => [([a], [a])] -> [([a], [a])]

ordpair :: (Ord (t a), Foldable t) => t a -> t a -> Maybe (t a, t a)

shortlex :: (Foldable t, Ord (t a)) => t a -> t a -> Ordering

knuthBendix2 :: Ord a => [([a], [a])] -> [([a], [a])]

merge :: Ord a => [a] -> [a] -> [a]

knuthBendix3 :: Ord a => [([a], [a])] -> [([a], [a])]

knuthBendix :: Ord a => [([a], [a])] -> [([a], [a])]

Implementation of the Knuth-Bendix algorithm. Given a list of relations, return a confluent rewrite system. The algorithm is not guaranteed to terminate.

nfs :: Ord a => ([a], [([a], [a])]) -> [[a]]

Given generators and a confluent rewrite system, return (normal forms of) all elements

elts :: Ord a => ([a], [([a], [a])]) -> [[a]]

Given generators and relations, return (normal forms of) all elements

newtype SGen

Constructors

S Int

s_ :: Int -> SGen

s1 :: SGen

s2 :: SGen

s3 :: SGen

_S :: Int -> ([SGen], [([SGen], [t])])

_S' :: Int -> ([SGen], [([SGen], [SGen])])

tri :: Int -> Int -> Int -> ([Char], [([Char], [Char])])

_D :: Int -> Int -> Int -> ([Char], [([Char], [Char])])