Math.Algebra.LinearAlgebra

Description

A module providing elementary operations involving scalars, vectors, and matrices over a ring or field. Vectors are represented as [a], matrices as [[a]]. (No distinction is made between row and column vectors.) It is the caller's responsibility to ensure that the lists have the correct number of elements.

The mnemonic for many of the arithmetic operations is that the number of angle brackets on each side indicates the dimension of the argument on that side. For example, v <*>> m is multiplication of a vector on the left by a matrix on the right.

Documentation

(<+>) :: Num a => [a] -> [a] -> [a] infixl 6

u <+> v returns the sum u+v of vectors

(<->) :: Num a => [a] -> [a] -> [a] infixl 6

u <-> v returns the difference u-v of vectors

(*>) :: Num a => a -> [a] -> [a] infixr 8

k *> v returns the product k*v of the scalar k and the vector v

(<.>) :: Num a => [a] -> [a] -> a infixl 7

u <.> v returns the dot product of vectors (also called inner or scalar product)

(<*>) :: Num a => [a] -> [a] -> [[a]] infixl 7

u <*> v returns the tensor product of vectors (also called outer or matrix product)

(<<+>>) :: Num a => [[a]] -> [[a]] -> [[a]] infixl 6

a <<+>> b returns the sum a+b of matrices

(<<->>) :: Num a => [[a]] -> [[a]] -> [[a]] infixl 6

a <<->> b returns the difference a-b of matrices

(<<*>>) :: Num a => [[a]] -> [[a]] -> [[a]] infixl 7

a <<*>> b returns the product a*b of matrices

(*>>) :: Num a => a -> [[a]] -> [[a]] infixr 8

k *>> m returns the product k*m of the scalar k and the matrix m

(<<*>) :: Num a => [[a]] -> [a] -> [a] infixr 7

m <<*> v is multiplication of a vector by a matrix on the left

(<*>>) :: Num a => [a] -> [[a]] -> [a] infixl 7

v <*>> m is multiplication of a vector by a matrix on the right

fMatrix :: (Enum t1, Num t1) => t1 -> (t1 -> t1 -> t) -> [[t]]

fMatrix' :: (Enum t1, Num t1) => t1 -> (t1 -> t1 -> t) -> [[t]]

idMx :: Num a => Int -> [[a]]

iMx :: Num t => Int -> [[t]]

iMx n is the n*n identity matrix

jMx :: Num t => Int -> [[t]]

jMx n is the n*n matrix of all 1s

zMx :: Num t => Int -> [[t]]

zMx n is the n*n matrix of all 0s

inverse :: (Eq a, Fractional a) => [[a]] -> Maybe [[a]]

The inverse of a matrix (over a field), if it exists

inverse1 :: (Fractional a, Eq a) => [[a]] -> [[a]]

inverse2 :: (Num a, Eq a) => [[a]] -> [[a]]

(!) :: [a] -> Int -> a

rowEchelonForm :: (Fractional a, Eq a) => [[a]] -> [[a]]

reducedRowEchelonForm :: (Eq a, Fractional a) => [[a]] -> [[a]]

solveLinearSystem :: (Eq a, Fractional a) => [[a]] -> [a] -> Maybe [a]

isZero :: (Num a, Eq a, Foldable t) => t a -> Bool

inSpanRE :: (Num a, Eq a) => [[a]] -> [a] -> Bool

rank :: (Fractional a, Eq a) => [[a]] -> Int

kernel :: (Fractional a, Ord a) => [[a]] -> [[a]]

kernelRRE :: (Num a, Ord a) => [[a]] -> [[a]]

det :: (Eq a, Fractional a) => [[a]] -> a

The determinant of a matrix (over a field)