Math.Algebra.NonCommutative.TensorAlgebra

Description

A module defining the tensor, symmetric, and exterior algebras. This module has been partially superceded by Math.Algebras.TensorAlgebra, which should be used in preference. This module is likely to be removed at some point.

Documentation

data Basis

Constructors

E Int

e_ :: Int -> NPoly Q Basis

e1 :: NPoly Q Basis

e2 :: NPoly Q Basis

e3 :: NPoly Q Basis

e4 :: NPoly Q Basis

dim :: NPoly t Basis -> Int

tensorBasis :: Int -> [NPoly Q Basis]

extRelations :: Int -> [NPoly Q Basis]

extnf :: NPoly Q Basis -> NPoly Q Basis

exteriorBasis :: Int -> [NPoly Q Basis]

symRelations :: Int -> [NPoly Q Basis]

symnf :: NPoly Q Basis -> NPoly Q Basis

symmetricBasis :: Int -> [NPoly Q Basis]

weylRelations :: Int -> [NPoly Q Basis]

weylnf :: Int -> NPoly Q Basis -> NPoly Q Basis

weylBasis :: Int -> [NPoly Q Basis]

data WeylGens

Constructors

X Int

D Int

d_ :: Int -> NPoly Q WeylGens

x_ :: Int -> NPoly Q WeylGens

d1 :: NPoly Q WeylGens

d2 :: NPoly Q WeylGens

d3 :: NPoly Q WeylGens

x1 :: NPoly Q WeylGens

x2 :: NPoly Q WeylGens

x3 :: NPoly Q WeylGens

comm :: Num a => a -> a -> a

delta :: (Num t, Eq a) => a -> a -> t

weylRelations' :: Int -> [NPoly Q WeylGens]

weylnf' :: NPoly Q WeylGens -> NPoly Q WeylGens

weylBasis' :: Int -> [NPoly Q WeylGens]