Math.Algebras.Commutative

Description

A module defining the algebra of commutative polynomials over a field k.

Most users should probably use Math.CommutativeAlgebra.Polynomial instead, which is basically the same thing but more fully-featured. This module will probably be deprecated at some point, but remains for now because it has a simpler implementation which may be more helpful for people wanting to understand the code.

Documentation

data GlexMonomial v

Constructors

Glex Int [(v, Int)]

type GlexPoly k v = Vect k (GlexMonomial v)

glexVar :: Num k => v -> GlexPoly k v

glexVar creates a variable in the algebra of commutative polynomials with Glex term ordering. For example, the following code creates variables called x, y and z:

[x,y,z] = map glexVar ["x","y","z"] :: GlexPoly Q String

class Monomial m where

Minimal complete definition

var, powers

Methods

var :: v -> Vect Q (m v)

powers :: m v -> [(v, Int)]

bind :: (Monomial m, Eq k, Num k, Ord b, Show b, Algebra k b) => Vect k (m v) -> (v -> Vect k b) -> Vect k b

In effect, we have (Num k, Monomial m) => Monad (v -> Vect k (m v)), with return = var, and (>>=) = bind. However, we can't express this directly in Haskell, firstly because of the Ord b constraint, secondly because Haskell doesn't support type functions.

lt :: Vect t1 t -> (t, t1)

class DivisionBasis b where

Minimal complete definition

dividesB, divB

Methods

dividesB :: b -> b -> Bool

divB :: b -> b -> b

dividesT :: DivisionBasis b => (b, t1) -> (b, t) -> Bool

divT :: (Fractional t, DivisionBasis t1) => (t1, t) -> (t1, t) -> (t1, t)

quotRemMP :: (Fractional t, DivisionBasis b, Algebra t b, Show b, Ord b, Eq t) => Vect t b -> [Vect t b] -> ([Vect t b], Vect t b)

(%%) :: (Eq k, Fractional k, Ord b, Show b, Algebra k b, DivisionBasis b) => Vect k b -> [Vect k b] -> Vect k b infixl 7

(%%) reduces a polynomial with respect to a list of polynomials.