Math.Algebras.NonCommutative

Description

A module defining the algebra of non-commutative polynomials over a field k

Documentation

data NonComMonomial v

Constructors

NCM Int [v]

class Monomial m where

Minimal complete definition

var, powers

Methods

var :: v -> Vect Q (m v)

powers :: Eq v => m v -> [(v, Int)]

bind :: (Eq t, Eq k, Monomial m, Algebra k b, Show b, Ord b, Num k) => Vect k (m t) -> (t -> Vect k b) -> Vect k b

type NCPoly v = Vect Q (NonComMonomial v)

class DivisionBasis m where

Minimal complete definition

divM

Methods

divM :: m -> m -> Maybe (m, m)

ncm :: [v] -> NonComMonomial v

lm :: Vect t t1 -> t1

lc :: Vect t1 t -> t1

lt :: Vect k b -> Vect k b

quotRemNP :: (Fractional t, DivisionBasis m, Algebra t m, Show m, Ord m, Eq t) => Vect t m -> [Vect t m] -> ([(Vect t m, Vect t m)], Vect t m)

remNP :: (Fractional t, DivisionBasis m, Algebra t m, Show m, Ord m, Eq t) => Vect t m -> [Vect t m] -> Vect t m

(%%) :: (Eq t, Ord m, Show m, Algebra t m, DivisionBasis m, Fractional t) => Vect t m -> [Vect t m] -> Vect t m infixl 7