Math.Algebras.Quaternions

Description

A module defining the algebra of quaternions over an arbitrary field.

The quaternions are the algebra defined by the basis {1,i,j,k}, where i^2 = j^2 = k^2 = ijk = -1

Documentation

data HBasis

Constructors

One

I

J

K

type Quaternion k = Vect k HBasis

i :: Num k => Quaternion k

The quaternions have {1,i,j,k} as basis, where i^2 = j^2 = k^2 = ijk = -1.

j :: Num k => Quaternion k

The quaternions have {1,i,j,k} as basis, where i^2 = j^2 = k^2 = ijk = -1.

k :: Num k => Quaternion k

The quaternions have {1,i,j,k} as basis, where i^2 = j^2 = k^2 = ijk = -1.

class Algebra k a => HasConjugation k a where

Minimal complete definition

conj, sqnorm

Methods

conj :: Vect k a -> Vect k a

A conjugation operation is required to satisfy the following laws:

conj (x+y) = conj x + conj y

conj (x*y) = conj y * conj x (note the order-reversal)

conj (conj x) = x

conj x = x if and only if x in k

sqnorm :: Vect k a -> k

The squared norm is defined as sqnorm x = x * conj x. It satisfies:

sqnorm (x*y) = sqnorm x * sqnorm y

sqnorm (unit k) = k^2, for k a scalar

scalarPart :: Num k => Quaternion k -> k

The scalar part of the quaternion w+xi+yj+zk is w. Also called the real part.

vectorPart :: (Eq k, Num k) => Quaternion k -> Quaternion k

The vector part of the quaternion w+xi+yj+zk is xi+yj+zk. Also called the pure part.

(<.>) :: (Eq k, Num k) => Vect k HBasis -> Vect k HBasis -> k

(^-) :: (Num a, Fractional a1, Eq a) => a1 -> a -> a1

refl :: (HasConjugation k a, Show a, Ord a, Eq k, Num k) => Vect k a -> Vect k a -> Vect k a

asMatrix :: (Num t, Eq t) => (Vect t HBasis -> Vect t HBasis) -> [Vect t HBasis] -> [[t]]

reprSO3' :: Fractional a => a -> a -> a

reprSO3 :: (Eq k, Fractional k) => Quaternion k -> [[k]]

Given a non-zero quaternion q in H, the map x -> q^-1 * x * q defines an action on the 3-dimensional vector space of pure quaternions X (ie linear combinations of i,j,k). It turns out that this action is a rotation of X, and this is a surjective group homomorphism from H* onto SO3. If we restrict q to the group of unit quaternions (those of norm 1), then this homomorphism is 2-to-1 (since q and -q give the same rotation). This shows that the multiplicative group of unit quaternions is isomorphic to Spin3, the double cover of SO3.

reprSO3 q returns the 3*3 matrix representing this map.

reprSO4' :: Fractional a => (a, a) -> a -> a

reprSO4 :: (Eq k, Fractional k) => (Quaternion k, Quaternion k) -> [[k]]

Given a pair of unit quaternions (l,r), the map x -> l^-1 * x * r defines an action on the 4-dimensional space of quaternions. It turns out that this action is a rotation, and this is a surjective group homomorphism onto SO4. The homomorphism is 2-to-1 (since (l,r) and (-l,-r) give the same map). This shows that the multiplicative group of pairs of unit quaternions (with pointwise multiplication) is isomorphic to Spin4, the double cover of SO4.

reprSO4 (l,r) returns the 4*4 matrix representing this map.

reprSO4d :: (Fractional k, Eq k) => Vect k (DSum HBasis HBasis) -> [[k]]

one' :: Num k => Vect k (Dual HBasis)

i' :: Num k => Vect k (Dual HBasis)

j' :: Num k => Vect k (Dual HBasis)

k' :: Num k => Vect k (Dual HBasis)

Orphan instances

(Eq k, Fractional k, Ord a, Show a, HasConjugation k a) => Fractional (Vect k a)
If an algebra has a conjugation operation, then it has multiplicative inverses, via 1/x = conj x / sqnorm x

Methods

(/) :: Vect k a -> Vect k a -> Vect k a

recip :: Vect k a -> Vect k a

fromRational :: Rational -> Vect k a