Math.Algebras.Structures

Description

A module defining various algebraic structures that can be defined on vector spaces - specifically algebra, coalgebra, bialgebra, Hopf algebra, module, comodule

Documentation

class Mon m where

Monoid

Minimal complete definition

munit, mmult

Methods

munit :: m

mmult :: m -> m -> m

class Algebra k b where

Caution: If we declare an instance Algebra k b, then we are saying that the vector space Vect k b is a k-algebra. In other words, we are saying that b is the basis for a k-algebra. So a more accurate name for this class would have been AlgebraBasis.

Minimal complete definition

unit, mult

Methods

unit :: k -> Vect k b

mult :: Vect k (Tensor b b) -> Vect k b

unit' :: (Eq k, Num k, Algebra k b) => Vect k () -> Vect k b

Sometimes it is more convenient to work with this version of unit.

class Coalgebra k b where

An instance declaration for Coalgebra k b is saying that the vector space Vect k b is a k-coalgebra.

Minimal complete definition

counit, comult

Methods

counit :: Vect k b -> k

comult :: Vect k b -> Vect k (Tensor b b)

counit' :: (Eq k, Num k, Coalgebra k b) => Vect k b -> Vect k ()

Sometimes it is more convenient to work with this version of counit.

class (Algebra k b, Coalgebra k b) => Bialgebra k b

A bialgebra is an algebra which is also a coalgebra, subject to the compatibility conditions that counit and comult must be algebra morphisms (or equivalently, that unit and mult must be coalgebra morphisms)

class Bialgebra k b => HopfAlgebra k b where

Minimal complete definition

antipode

Methods

antipode :: Vect k b -> Vect k b

newtype SetCoalgebra b

Constructors

SC b

newtype MonoidCoalgebra m

Constructors

MC m

class Algebra k a => Module k a m where

Minimal complete definition

action

Methods

action :: Vect k (Tensor a m) -> Vect k m

(*.) :: (Num k, Module k a m) => Vect k a -> Vect k m -> Vect k m

class Coalgebra k c => Comodule k c n where

Minimal complete definition

coaction

Methods

coaction :: Vect k n -> Vect k (Tensor c n)

class HasPairing k u v where

A pairing is a non-degenerate bilinear form U x V -> k. We are typically interested in pairings having additional properties. For example:

A bialgebra pairing is a pairing between bialgebras A and B such that the mult in A is adjoint to the comult in B, and vice versa, and the unit in A is adjoint to the counit in B, and vice versa.

A Hopf pairing is a bialgebra pairing between Hopf algebras A and B such that the antipodes in A and B are adjoint.

Minimal complete definition

pairing

Methods

pairing :: Vect k (Tensor u v) -> Vect k ()

pairing' :: (Num k, HasPairing k u v) => Vect k u -> Vect k v -> k

The pairing function with a more Haskellish type signature

Orphan instances

(Eq k, Num k, Eq b, Ord b, Show b, Algebra k b) => Num (Vect k b)

Methods

(+) :: Vect k b -> Vect k b -> Vect k b

(-) :: Vect k b -> Vect k b -> Vect k b

(*) :: Vect k b -> Vect k b -> Vect k b

negate :: Vect k b -> Vect k b

abs :: Vect k b -> Vect k b

signum :: Vect k b -> Vect k b

fromInteger :: Integer -> Vect k b