Math.Algebras.VectorSpace

Description

A module defining the type and operations of free k-vector spaces over a basis b (for a field k)

Documentation

newtype Vect k b

Given a field type k and a basis type b, Vect k b is the type of the free k-vector space over b. Elements (values) of Vect k b consist of k-linear combinations of elements (values) of b.

In order for Vect k b to be a vector space, it is necessary that k is a field (that is, an instance of Fractional). In practice, we often relax this condition, and require that k is a ring (that is, an instance of Num). In that case, Vect k b should more correctly be called (the type of) the free k-module over b.

Most of the code requires that b is an instance of Ord. This is primarily to enable us to simplify to a normal form.

Constructors

V [(b, k)]

coeff :: (Num k, Eq b) => b -> Vect k b -> k

Return the coefficient of the specified basis element in a vector

removeTerm :: (Eq k, Num k, Ord b) => b -> Vect k b -> Vect k b

Remove the term for a specified basis element from a vector

zerov :: Vect k b

The zero vector

add :: (Eq k, Num k, Ord b) => Vect k b -> Vect k b -> Vect k b

Addition of vectors

(<+>) :: (Eq k, Num k, Ord b) => Vect k b -> Vect k b -> Vect k b infixl 6

Addition of vectors (same as add)

sumv :: (Eq k, Num k, Ord b) => [Vect k b] -> Vect k b

Sum of a list of vectors

negatev :: (Eq k, Num k) => Vect k b -> Vect k b

Negation of a vector

(<->) :: (Eq k, Num k, Ord b) => Vect k b -> Vect k b -> Vect k b infixl 6

Subtraction of vectors

smultL :: (Eq k, Num k) => k -> Vect k b -> Vect k b

Scalar multiplication (on the left)

(*>) :: (Eq k, Num k) => k -> Vect k b -> Vect k b infixr 7

Same as smultL. Mnemonic is "multiply through (from the left)"

smultR :: (Eq k, Num k) => Vect k b -> k -> Vect k b

Scalar multiplication on the right

(<*) :: (Eq k, Num k) => Vect k b -> k -> Vect k b infixl 7

Same as smultR. Mnemonic is "multiply through (from the right)"

nf :: (Eq k, Num k, Ord b) => Vect k b -> Vect k b

Convert an element of Vect k b into normal form. Normal form consists in having the basis elements in ascending order, with no duplicates, and all coefficients non-zero

linear :: (Eq k, Num k, Ord b) => (a -> Vect k b) -> Vect k a -> Vect k b

A linear map between vector spaces A and B can be defined by giving its action on the basis elements of A. The action on all elements of A then follows by linearity.

If we have A = Vect k a, B = Vect k b, and f :: a -> Vect k b is a function from the basis elements of A into B, then linear f is the linear map that this defines by linearity.

type Trivial k = Vect k ()

Trivial k is the field k considered as a k-vector space. In maths, we would not normally make a distinction here, but in the code, we need this if we want to be able to put k as one side of a tensor product.

wrap :: (Eq k, Num k) => k -> Vect k ()

Wrap an element of the field k to an element of the trivial k-vector space

unwrap :: Num k => Vect k () -> k

Unwrap an element of the trivial k-vector space to an element of the field k

newtype Dual b

Given a finite vector space basis b, Dual b can be used to represent a basis for the dual vector space. The intention is that for a given individual basis element b_i, (Dual b_i) represents the indicator function for b_i, which takes b_i to 1 and all other basis elements to 0.

(Note that if the basis b is infinite, then Dual b may only represent a sub-basis of the dual vector space.)

Constructors

Dual b