Math.Combinatorics.CombinatorialHopfAlgebra

Description

A module defining the following Combinatorial Hopf Algebras, together with coalgebra or Hopf algebra morphisms between them:

Sh, the Shuffle Hopf algebra

SSym, the Malvenuto-Reutnenauer Hopf algebra of permutations

YSym, the (dual of the) Loday-Ronco Hopf algebra of binary trees

QSym, the Hopf algebra of quasi-symmetric functions (having a basis indexed by compositions)

Sym, the Hopf algebra of symmetric functions (having a basis indexed by integer partitions)

NSym, the Hopf algebra of non-commutative symmetric functions

Documentation

newtype Shuffle a

A basis for the shuffle algebra. As a vector space, the shuffle algebra is identical to the tensor algebra. However, we consider a different algebra structure, based on the shuffle product. Together with the deconcatenation coproduct, this leads to a Hopf algebra structure.

Constructors

Sh [a]

sh :: [a] -> Vect Q (Shuffle a)

Construct a basis element of the shuffle algebra

shuffles :: [a] -> [a] -> [[a]]

deconcatenations :: [a] -> [([a], [a])]

newtype SSymF

The fundamental basis for the Malvenuto-Reutenauer Hopf algebra of permutations, SSym.

Constructors

SSymF [Int]

ssymF :: [Int] -> Vect Q SSymF

Construct a fundamental basis element in SSym. The list of ints must be a permutation of [1..n], eg [1,2], [3,4,2,1].

shiftedConcat :: SSymF -> SSymF -> SSymF

prop_Associative :: Eq a => (a -> a -> a) -> (a, a, a) -> Bool

flatten :: (Ord a, Num t, Enum t) => [a] -> [t]

newtype SSymM

An alternative "monomial" basis for the Malvenuto-Reutenauer Hopf algebra of permutations, SSym. This basis is related to the fundamental basis by Mobius inversion in the poset of permutations with the weak order.

Constructors

SSymM [Int]

ssymM :: [Int] -> Vect Q SSymM

Construct a monomial basis element in SSym. The list of ints must be a permutation of [1..n], eg [1,2], [3,4,2,1].

inversions :: (Ord a, Num t, Enum t) => [a] -> [(t, t)]

weakOrder :: (Ord a, Ord a1) => [a1] -> [a] -> Bool

mu :: (Eq t1, Num t) => ([t1], t1 -> t1 -> Bool) -> t1 -> t1 -> t

ssymMtoF :: (Eq k, Num k) => Vect k SSymM -> Vect k SSymF

Convert an element of SSym represented in the monomial basis to the fundamental basis

ssymFtoM :: (Eq k, Num k) => Vect k SSymF -> Vect k SSymM

Convert an element of SSym represented in the fundamental basis to the monomial basis

ssymFtoDual :: (Eq k, Num k) => Vect k SSymF -> Vect k (Dual SSymF)

The isomorphism from SSym to its dual that takes a permutation in the fundamental basis to its inverse in the dual basis

data PBT a

A type for (rooted) planar binary trees. The basis elements of the Loday-Ronco Hopf algebra are indexed by these.

Although the trees are labelled, we're really only interested in the shapes of the trees, and hence in the type PBT (). The Algebra, Coalgebra and HopfAlgebra instances all ignore the labels. However, it is convenient to allow labels, as they can be useful for seeing what is going on, and they also make it possible to define various ways to create trees from lists of labels.

Constructors

T (PBT a) a (PBT a)

E

newtype YSymF a

The fundamental basis for (the dual of) the Loday-Ronco Hopf algebra of binary trees, YSym.

Constructors

YSymF (PBT a)

ysymF :: PBT a -> Vect Q (YSymF a)

Construct the element of YSym in the fundamental basis indexed by the given tree

nodecount :: Num t => PBT t1 -> t

leafcount :: Num t => PBT t1 -> t

prefix :: PBT t -> [t]

shapeSignature :: Num t => PBT t1 -> [t]

nodeCountTree :: Num a => PBT t -> PBT a

leafCountTree :: Num a => PBT t -> PBT a

lrCountTree :: Num t => PBT t1 -> PBT (t, t)

shape :: PBT a -> PBT ()

numbered :: Num a => PBT t -> PBT a

splits :: PBT a -> [(PBT a, PBT a)]

multisplits :: (Num t, Eq t) => t -> PBT a -> [[PBT a]]

graft :: [PBT a] -> PBT a -> PBT a

newtype YSymM

An alternative "monomial" basis for (the dual of) the Loday-Ronco Hopf algebra of binary trees, YSym.

Constructors

YSymM (PBT ())

ysymM :: PBT () -> Vect Q YSymM

Construct the element of YSym in the monomial basis indexed by the given tree

trees :: Int -> [PBT ()]

List all trees with the given number of nodes

tamariCovers :: PBT a -> [PBT a]

The covering relation for the Tamari partial order on binary trees

tamariUpSet :: Ord a => PBT a -> [PBT a]

The up-set of a binary tree in the Tamari partial order

tamariOrder :: PBT a -> PBT a -> Bool

The Tamari partial order on binary trees. This is only defined between trees of the same size (number of nodes). The result between trees of different sizes is undefined (we don't check).

ysymMtoF :: (Eq k, Num k) => Vect k YSymM -> Vect k (YSymF ())

Convert an element of YSym represented in the monomial basis to the fundamental basis

ysymFtoM :: (Eq k, Num k) => Vect k (YSymF ()) -> Vect k YSymM

Convert an element of YSym represented in the fundamental basis to the monomial basis

compositions :: Int -> [[Int]]

List the compositions of an integer n. For example, the compositions of 4 are [[1,1,1,1],[1,1,2],[1,2,1],[1,3],[2,1,1],[2,2],[3,1],[4]]

quasiShuffles :: [Int] -> [Int] -> [[Int]]

newtype QSymM

A type for the monomial basis for the quasi-symmetric functions, indexed by compositions.

Constructors

QSymM [Int]

qsymM :: [Int] -> Vect Q QSymM

Construct the element of QSym in the monomial basis indexed by the given composition

coarsenings :: Num a => [a] -> [[a]]

refinements :: [Int] -> [[Int]]

newtype QSymF

A type for the fundamental basis for the quasi-symmetric functions, indexed by compositions.

Constructors

QSymF [Int]

qsymF :: [Int] -> Vect Q QSymF

Construct the element of QSym in the fundamental basis indexed by the given composition

qsymMtoF :: (Eq k, Num k) => Vect k QSymM -> Vect k QSymF

Convert an element of QSym represented in the monomial basis to the fundamental basis

qsymFtoM :: (Eq k, Num k) => Vect k QSymF -> Vect k QSymM

Convert an element of QSym represented in the fundamental basis to the monomial basis

qsymPoly :: Int -> [Int] -> GlexPoly Q String

qsymPoly n is is the quasi-symmetric polynomial in n variables for the indices is. (This corresponds to the monomial basis for QSym.) For example, qsymPoly 3 [2,1] == x1^2*x2+x1^2*x3+x2^2*x3.

newtype SymM

A type for the monomial basis for Sym, the Hopf algebra of symmetric functions, indexed by integer partitions

Constructors

SymM [Int]

symM :: [Int] -> Vect Q SymM

Construct the element of Sym in the monomial basis indexed by the given integer partition

compositionsFromPartition :: Eq a => [a] -> [[a]]

symMult :: [Int] -> [Int] -> [[Int]]

newtype SymE

The elementary basis for Sym, the Hopf algebra of symmetric functions. Defined informally as > symE [n] = symM (replicate n 1) > symE lambda = product [symE [p] | p <- lambda]

Constructors

SymE [Int]

symE :: [Int] -> Vect Q SymE

symEtoM :: (Eq k, Num k) => Vect k SymE -> Vect k SymM

Convert from the elementary to the monomial basis of Sym

newtype SymH

The complete basis for Sym, the Hopf algebra of symmetric functions. Defined informally as > symH [n] = sum [symM lambda | lambda <- integerPartitions n] -- == all monomials of weight n > symH lambda = product [symH [p] | p <- lambda]

Constructors

SymH [Int]

symH :: [Int] -> Vect Q SymH

symHtoM :: (Eq k, Num k) => Vect k SymH -> Vect k SymM

Convert from the complete to the monomial basis of Sym

newtype NSym

A basis for NSym, the Hopf algebra of non-commutative symmetric functions, indexed by compositions

Constructors

NSym [Int]

nsym :: [Int] -> Vect Q NSym

descendingTree :: Ord a => [a] -> PBT a

descendingTreeMap :: (Eq k, Num k) => Vect k SSymF -> Vect k (YSymF ())

Given a permutation p of [1..n], we can construct a tree (the descending tree of p) as follows:

Split the permutation as p = ls ++ [n] ++ rs

Place n at the root of the tree, and recursively place the descending trees of ls and rs as the left and right children of the root

To bottom out the recursion, the descending tree of the empty permutation is of course the empty tree

This map between bases SSymF -> YSymF turns out to induce a morphism of Hopf algebras.

minPerm :: Num t => PBT t1 -> [t]

maxPerm :: Num t => PBT t1 -> [t]

leftLeafComposition :: PBT t -> [Int]

leftLeafComposition' :: YSymF t -> QSymF

leftLeafCompositionMap :: (Eq k, Num k) => Vect k (YSymF a) -> Vect k QSymF

A Hopf algebra morphism from YSymF to QSymF

descents :: Ord a => [a] -> [Int]

descentComposition :: (Ord a, Num t) => [a] -> [t]

descentMap :: (Eq k, Num k) => Vect k SSymF -> Vect k QSymF

Given a permutation of [1..n], its descents are those positions where the next number is less than the previous number. For example, the permutation [2,3,5,1,6,4] has descents from 5 to 1 and from 6 to 4. The descents can be regarded as cutting the permutation sequence into segments - 235-16-4 - and by counting the lengths of the segments, we get a composition 3+2+1. This map between bases SSymF -> QSymF turns out to induce a morphism of Hopf algebras.

underComposition :: QSymF -> SSymF

under :: PBT a -> PBT a -> PBT a

isUnderIrreducible :: PBT t -> Bool

underDecomposition :: PBT a -> [PBT a]

ysymmToSh :: Functor f => f YSymM -> f (Shuffle (PBT ()))

symToQSymM :: (Eq k, Num k) => Vect k SymM -> Vect k QSymM

The injection of Sym into QSym (defined over the monomial basis)

nsymToSymH :: (Eq k, Num k) => Vect k NSym -> Vect k SymH

A surjection of NSym onto Sym (defined over the complete basis)

nsymToSSym :: (Num k, Eq k) => Vect k NSym -> Vect k SSymF