Math.Combinatorics.Hypergraph

Description

A module defining a type for hypergraphs.

Documentation

data Hypergraph a

Constructors

H [a] [[a]]

hypergraph :: Ord a => [a] -> [[a]] -> Hypergraph a

toHypergraph :: Ord a => [a] -> [[a]] -> Hypergraph a

isUniform :: Ord a => Hypergraph a -> Bool

Is this hypergraph uniform - meaning that all blocks are of the same size

same :: Eq a => [a] -> Bool

fromGraph :: Graph a -> Hypergraph a

fromDesign :: Ord a => Design a -> Hypergraph a

incidenceGraph :: Ord a => Hypergraph a -> Graph (Either a [a])

incidenceMatrix :: (Num t, Eq a) => Hypergraph a -> [[t]]

fromIncidenceMatrix :: (Eq a1, Num a1, Num a, Ord a, Enum a) => [[a1]] -> Hypergraph a

isPartialLinearSpace :: Ord a => Hypergraph a -> Bool

isProjectivePlane :: Ord a => Hypergraph a -> Bool

Is this hypergraph a projective plane - meaning that any two lines meet in a unique point, and any two points lie on a unique line

isProjectivePlaneTri :: Ord a => Hypergraph a -> Bool

Is this hypergraph a projective plane with a triangle. This is a weak non-degeneracy condition, which eliminates all points on the same line, or all lines through the same point.

isProjectivePlaneQuad :: Ord a => Hypergraph a -> Bool

Is this hypergraph a projective plane with a quadrangle. This is a stronger non-degeneracy condition.

isGeneralizedQuadrangle :: Ord a => Hypergraph a -> Bool

grid :: (Enum t1, Enum t, Num t1, Num t, Ord t, Ord t1) => t1 -> t -> Hypergraph (t1, t)

dualGrid :: Integral a => a -> a -> Hypergraph a

isGenQuadrangle' :: Ord a => Hypergraph a -> Bool

isConfiguration :: Ord a => Hypergraph a -> Bool

Is this hypergraph a (projective) configuration.

fanoPlane :: Hypergraph Integer

heawoodGraph :: Graph (Either Integer [Integer])

The Heawood graph is the incidence graph of the Fano plane

desarguesConfiguration :: Hypergraph [Integer]

desarguesGraph :: Graph (Either [Integer] [[Integer]])

pappusConfiguration :: Hypergraph Integer

pappusGraph :: Graph (Either Integer [Integer])

coxeterGraph :: Graph [Integer]

duads :: [[Integer]]

synthemes :: [[[Integer]]]

tutteCoxeterGraph :: Graph (Either [Integer] [[Integer]])

The Tutte-Coxeter graph, also called the Tutte 8-cage

intersectionGraph :: Ord t => Hypergraph t -> Graph [t]