Math.Combinatorics.IncidenceAlgebra

Documentation

data Interval a

A type to represent an interval in a poset. The (closed) interval [x,y] is the set {z | x <= z <= y} within the poset. Note that the "empty interval" is not an interval - that is, the interval [x,y] is only defined for x <= y. The (closed) intervals within a poset form a basis for the incidence algebra as a k-vector space.

Constructors

Iv (Poset a) (a, a)

ivPoset :: Interval t -> Poset t

intervalIsos :: (Ord b, Ord a) => Interval a -> Interval b -> [[(a, b)]]

isIntervalIso :: (Ord b, Ord a) => Interval a -> Interval b -> Bool

intervalIsoMap :: Ord a => Poset a -> Map (Interval a) (Maybe (Interval a))

intervalIsoClasses :: Ord a => Poset a -> [Interval a]

List representatives of the order isomorphism classes of intervals in a poset

unitIA :: (Eq k, Num k, Ord a) => Poset a -> Vect k (Interval a)

The unit of the incidence algebra of a poset

basisIA :: Num k => Poset a -> [Vect k (Interval a)]

zetaIA :: (Eq k, Num k, Ord a) => Poset a -> Vect k (Interval a)

The zeta function of a poset

muIA1 :: (Show a, Ord a, Eq k, Num k) => Poset a -> Vect k (Interval a)

muIA :: (Eq k, Num k, Ord a) => Poset a -> Vect k (Interval a)

The Mobius function of a poset

invIA1 :: (Ord t, Fractional a, Eq a) => Vect a (Interval t) -> Vect a (Interval t)

invIA :: (Eq k, Fractional k, Ord a) => Vect k (Interval a) -> Maybe (Vect k (Interval a))

The inverse of an element in the incidence algebra of a poset. This is only defined for elements which are non-zero on all intervals (x,x)

numChainsIA :: (Ord a, Show a) => Poset a -> Vect Q (Interval a)

A function (ie element of the incidence algebra) that counts the total number of chains in each interval

etaIA :: (Eq k, Ord a, Num k) => Poset a -> Vect k (Interval a)

numMaximalChainsIA :: (Ord a, Show a) => Poset a -> Vect Q (Interval a)

A function (ie element of the incidence algebra) that counts the number of maximal chains in each interval

muC :: (Num k, Eq k) => Int -> Vect k (Interval Int)

muB :: (Num k, Eq k) => Int -> Vect k (Interval [Int])

muL :: (Num a, Ord a) => Int -> [a] -> Vect Int (Interval [[a]])

toIsoClasses :: (Eq k, Num k, Ord a) => Vect k (Interval a) -> Vect k (Interval a)

toIsoClasses is the linear map from the incidence Hopf algebra of a poset to itself, in which each interval is mapped to (the minimal representative of) its isomorphism class. Thus the result can be considered as a linear combination of isomorphism classes of intervals, rather than of intervals themselves. Note that if this operation is to be performed repeatedly for the same poset, then it is more efficient to use toIsoClasses' poset, which memoizes the isomorphism class lookup table.

toIsoClasses' :: (Eq k, Num k, Ord a) => Poset a -> Vect k (Interval a) -> Vect k (Interval a)

Given a poset, toIsoClasses' poset is the linear map from the incidence Hopf algebra of the poset to itself, in which each interval is mapped to (the minimal representative of) its isomorphism class.