Math.Combinatorics.StronglyRegularGraph

Description

A module defining various strongly regular graphs, including the Clebsch, Hoffman-Singleton, Higman-Sims, and McLaughlin graphs.

A strongly regular graph with parameters (n,k,lambda,mu) is a (simple) graph with n vertices, in which the number of common neighbours of x and y is k, lambda or mu according as whether x and y are equal, adjacent, or non-adjacent. (In particular, it is a k-regular graph.)

Strongly regular graphs are highly symmetric, and have large automorphism groups.

Documentation

srgParams :: Ord t => Graph t -> Maybe (Int, Int, Int, Int)

isSRG :: Ord t => Graph t -> Bool

t' :: (Num a, Num t, Enum a, Enum t, Ord t, Ord a) => a -> Graph t

t :: (Enum a, Num a, Ord a) => a -> Graph [a]

l2' :: (Num a, Num t, Enum a, Enum t, Ord t, Ord a) => a -> Graph t

l2 :: (Enum a, Num a, Ord a) => a -> Graph (a, a)

paleyGraph :: (Num t, Ord t) => [t] -> Graph t

clebsch' :: Graph Integer

clebsch :: Graph [Integer]

clebsch2 :: Graph DesignVertex

triples :: [[Integer]]

heptads :: [[[Integer]]]

(+^) :: Ord a => [[a]] -> Permutation a -> [[a]]

(+^^) :: Ord a => [[a]] -> [Permutation a] -> [[[a]]]

hoffmanSingleton' :: Graph Integer

hoffmanSingleton :: Graph (Either [[Integer]] [Integer])

inducedA7 :: Permutation Integer -> Permutation (Either [[Integer]] [Integer])

hsA7 :: [Permutation Integer]

gewirtz' :: Graph Integer

gewirtz :: Graph [Integer]

data DesignVertex

Constructors

C

P Integer

B [Integer]

higmanSimsGraph' :: Graph Integer

higmanSimsGraph :: Graph DesignVertex

inducedM22 :: Permutation Integer -> Permutation DesignVertex

higmanSimsM22 :: [Permutation Integer]

_HS2 :: [Permutation DesignVertex]

_HS :: [Permutation DesignVertex]

sp2 :: Int -> Graph [F2]

sp :: Int -> Graph [F2]

switch :: Ord t => Graph t -> [t] -> Graph t

schlafli' :: Graph Integer

schlafli :: Graph Integer

mcLaughlin' :: Graph Integer

mcLaughlin :: Graph DesignVertex

_McL2 :: [Permutation DesignVertex]

_McL :: [Permutation DesignVertex]