Math.CommutativeAlgebra.GroebnerBasis

Description

A module providing an efficient implementation of the Buchberger algorithm for calculating the (reduced) Groebner basis for an ideal, together with some straightforward applications.

Documentation

sPoly :: (Monomial b, Fractional k, Algebra k b, Ord b, Eq k) => Vect k b -> Vect k b -> Vect k b

isGB :: (Algebra k b, Ord b, Monomial b, Fractional k, Eq k) => [Vect k b] -> Bool

gb1 :: (Eq k, Ord b, Algebra k b, Monomial b, Fractional k) => [Vect k b] -> [Vect k b]

pairWith :: (a1 -> a1 -> a) -> [a1] -> [a]

reduce :: (Algebra k b, Fractional k, Ord b, Ord k, Monomial b) => [Vect k b] -> [Vect k b]

gb2 :: (Monomial b, Ord k, Ord b, Fractional k, Algebra k b) => [Vect k b] -> [Vect k b]

(!) :: [a] -> Int -> a

gb2a :: (Monomial b, Ord k, Ord b, Fractional k, Algebra k b) => [Vect k b] -> [Vect k b]

gb3 :: (Monomial b, Ord k, Ord b, Fractional k, Algebra k b) => [Vect k b] -> [Vect k b]

gb4 :: (Monomial b, Ord k, Ord b, Fractional k, Algebra k b) => [Vect k b] -> [Vect k b]

mergeBy :: (a -> a -> Ordering) -> [a] -> [a] -> [a]

gb :: (Fractional k, Ord k, Monomial m, Ord m, Algebra k m) => [Vect k m] -> [Vect k m]

Given a list of polynomials over a field, return a Groebner basis for the ideal generated by the polynomials.

sugar :: (Monomial m, Monomial m2, Monomial m1) => Vect b1 m2 -> Vect b m -> m1 -> Int

cmpNormal :: (Ord t2, Ord t4) => ((t5, t4), (t3, t2)) -> ((t1, t4), (t, t2)) -> Ordering

cmpSug :: (Num t4, Ord t1, Ord t3, Ord t4) => ((t4, t3), (t2, t1)) -> ((t4, t3), (t, t1)) -> Ordering

memberGB :: (Algebra k m, Ord m, Monomial m, Fractional k, Eq k) => Vect k m -> [Vect k m] -> Bool

memberI :: (Fractional k, Ord k, Monomial m, Ord m, Algebra k m) => Vect k m -> [Vect k m] -> Bool

memberI f gs returns whether f is in the ideal generated by gs

sumI :: (Fractional k, Ord k, Monomial m, Ord m, Algebra k m) => [Vect k m] -> [Vect k m] -> [Vect k m]

Given ideals I and J, their sum is defined as I+J = {f+g | f <- I, g <- J}.

If fs and gs are generators for I and J, then sumI fs gs returns generators for I+J.

The geometric interpretation is that the variety of the sum is the intersection of the varieties, ie V(I+J) = V(I) intersect V(J)

productI :: (Fractional k, Ord k, Monomial m, Ord m, Algebra k m) => [Vect k m] -> [Vect k m] -> [Vect k m]

Given ideals I and J, their product I.J is the ideal generated by all products {f.g | f <- I, g <- J}.

If fs and gs are generators for I and J, then productI fs gs returns generators for I.J.

The geometric interpretation is that the variety of the product is the union of the varieties, ie V(I.J) = V(I) union V(J)

intersectI :: (Fractional k, Ord k, Monomial m, Ord m) => [Vect k m] -> [Vect k m] -> [Vect k m]

The intersection of ideals I and J is the set of all polynomials which belong to both I and J.

If fs and gs are generators for I and J, then intersectI fs gs returns generators for the intersection of I and J

The geometric interpretation is that the variety of the intersection is the union of the varieties, ie V(I intersect J) = V(I) union V(J).

The reason for prefering the intersection over the product is that the intersection of radical ideals is radical, whereas the product need not be.

toElimFst :: (Mon b, Functor f) => f a -> f (Elim2 a b)

toElimSnd :: (Mon a, Functor f) => f b -> f (Elim2 a b)

isElimFst :: (Mon a, Eq a) => Vect b (Elim2 a t) -> Bool

fromElimSnd :: Functor f => f (Elim2 t b) -> f b

eliminateFst :: (Ord b, Ord t, Ord b1, Monomial b, Monomial t, Fractional b1) => [Vect b1 (Elim2 t b)] -> [Vect b1 b]

quotientI :: (Fractional k, Ord k, Monomial m, Ord m, Algebra k m) => [Vect k m] -> [Vect k m] -> [Vect k m]

Given ideals I and J, their quotient is defined as I:J = {f | f <- R, f.g is in I for all g in J}.

If fs and gs are generators for I and J, then quotientI fs gs returns generators for I:J.

The ideal quotient is the algebraic analogue of the Zariski closure of a difference of varieties. V(I:J) contains the Zariski closure of V(I)-V(J), with equality if k is algebraically closed and I is a radical ideal.

quotientP :: (Ord k, Ord m, Fractional k, Monomial m, Algebra k m) => [Vect k m] -> Vect k m -> [Vect k m]

eliminate :: (Eq k, Fractional k, Ord k, MonomialConstructor m, Monomial (m v), Ord (m v)) => [Vect k (m v)] -> [Vect k (m v)] -> [Vect k (m v)]

eliminate vs gs returns the elimination ideal obtained from the ideal generated by gs by eliminating the variables vs.

mbasis :: (Ord a, Num a) => [a] -> [a]

mbasisQA :: (Fractional k, Ord k, Monomial m, Ord m, Algebra k m) => [Vect k m] -> [Vect k m] -> [Vect k m]

Given variables vs, and a Groebner basis gs, mbasisQA vs gs returns a monomial basis for the quotient algebra k[vs]/<gs>. For example, mbasisQA [x,y] [x^2+y^2-1] returns a monomial basis for k[x,y]/<x^2+y^2-1>. In general, the monomial basis is likely to be infinite.

ltIdeal :: (Fractional k, Ord k, Monomial m, Ord m, Algebra k m) => [Vect k m] -> [Vect k m]

Given an ideal I, the leading term ideal lt(I) consists of the leading terms of all elements of I. If I is generated by gs, then ltIdeal gs returns generators for lt(I).

numMonomials :: Integral a => a -> a -> Integer

hilbertFunQA :: (Fractional k, Ord k, Monomial m, Ord m, Algebra k m) => [Vect k m] -> [Vect k m] -> Int -> Integer

Given variables vs, and a homogeneous ideal gs, hilbertFunQA vs gs returns the Hilbert function for the quotient algebra k[vs]/<gs>. Given an integer i, the Hilbert function returns the number of degree i monomials in a basis for k[vs]/<gs>. For a homogeneous ideal, this number is independent of the monomial ordering used (even though the elements of the monomial basis themselves are dependent on the ordering).

If the ideal I is not homogeneous, then R/I is not graded, and the Hilbert function is not well-defined. Specifically, the number of degree i monomials in a basis is likely to depend on which monomial ordering you use.

hilbertSeriesQA1 :: (Monomial m, Fractional k, Ord k, Ord m, Algebra k m) => [Vect k m] -> [Vect k m] -> [Int]

hilbertSeriesQA :: (Fractional k, Ord k, Monomial m, Ord m, Algebra k m) => [Vect k m] -> [Vect k m] -> [Integer]

Given variables vs, and a homogeneous ideal gs, hilbertSeriesQA vs gs returns the Hilbert series for the quotient algebra k[vs]/<gs>. The Hilbert series should be interpreted as a formal power series where the coefficient of t^i is the Hilbert function evaluated at i. That is, the i'th element in the series is the number of degree i monomials in a basis for k[vs]/<gs>.

hilbertSeriesQA' :: (Fractional k, Ord k, MonomialConstructor m, Ord (m v), Monomial (m v), Algebra k (m v)) => [Vect k (m v)] -> [Integer]

In the case where every variable v occurs in some generator g of the homogeneous ideal (the usual case), then the vs can be inferred from the gs. hilbertSeriesQA' gs returns the Hilbert series for the quotient algebra k[vs]/<gs>.

hilbertPolyQA :: (Fractional k, Ord k, Monomial m, Ord m, Algebra k m) => [Vect k m] -> [Vect k m] -> GlexPoly Q String

For i >> 0, the Hilbert function becomes a polynomial in i, called the Hilbert polynomial.

hilbertPolyQA' :: (Fractional k, Ord k, MonomialConstructor m, Ord (m v), Monomial (m v), Algebra k (m v)) => [Vect k (m v)] -> GlexPoly Q String

dim :: (Algebra k m, Ord m, Ord k, Monomial m, Fractional k) => [Vect k m] -> [Vect k m] -> Int

dim' :: (Algebra k (m v), Monomial (m v), Ord (m v), Ord k, MonomialConstructor m, Fractional k) => [Vect k (m v)] -> Int