Math.CommutativeAlgebra.Polynomial

Description

A module defining the algebra of commutative polynomials over a field k. Polynomials are represented as the free k-vector space with the monomials as basis.

A monomial ordering is required to specify how monomials are to be ordered. The Lex, Glex, and Grevlex monomial orders are defined, with the possibility to add others.

In order to make use of this module, some variables must be defined, for example:

[t,u,v,x,y,z] = map glexvar ["t","u","v","x","y","z"]

Documentation

class (Eq m, Show m, Mon m) => Monomial m where

In order to work with monomials, we need to be able to multiply them and divide them. Multiplication is defined by the Mon (monoid) class. Division is defined in this class. The functions here are primarily intended for internal use only.

Minimal complete definition

mdivides, mdiv, mgcd, mlcm, mcoprime, mdeg

Methods

mdivides :: m -> m -> Bool

mdiv :: m -> m -> m

mgcd :: m -> m -> m

mlcm :: m -> m -> m

mcoprime :: m -> m -> Bool

mdeg :: m -> Int

mproperlydivides :: Monomial m => m -> m -> Bool

class MonomialConstructor m where

We want to be able to construct monomials over any set of variables that we choose. Although we will often use String as the type of our variables, it is useful to define polymorphic types for monomials.

Minimal complete definition

mvar, mindices

Methods

mvar :: v -> m v

mindices :: m v -> [(v, Int)]

var :: (Num k, MonomialConstructor m) => v -> Vect k (m v)

var v creates a variable in the vector space of polynomials. For example, if we want to work in Q[x,y,z], we might define:

[x,y,z] = map var ["x","y","z"] :: [GlexPoly Q String]

Notice that, in general, it is necessary to provide a type annotation so that the compiler knows which field and which term order to use.

data MonImpl v

The underlying implementation of monomials in variables of type v. Most often, we will be interested in MonImpl String, with the variable "x" represented by M 1 [("x",1)]. However, other types can be used instead.

No Ord instance is defined for MonImpl v, so it cannot be used as the basis for a free vector space of polynomials. Instead, several different newtype wrappers are provided, corresponding to different monomial orderings.

Constructors

M Int [(v, Int)]

newtype Lex v

A type representing monomials with Lex ordering.

Lex stands for lexicographic ordering. For example, in Lex ordering, monomials up to degree two would be ordered as follows: x^2+xy+xz+x+y^2+yz+y+z^2+z+1.

Constructors

Lex (MonImpl v)

type LexPoly k v = Vect k (Lex v)

A type representing polynomials with Lex term ordering.

lexvar :: v -> LexPoly Q v

lexvar v creates a variable in the algebra of commutative polynomials over Q with Lex term ordering. It is provided as a shortcut, to avoid having to provide a type annotation, as with var. For example, the following code creates variables called x, y and z:

[x,y,z] = map lexvar ["x","y","z"]

newtype Glex v

A type representing monomials with Glex ordering.

Glex stands for graded lexicographic. Thus monomials are ordered first by degree, then by lexicographic order. For example, in Glex ordering, monomials up to degree two would be ordered as follows: x^2+xy+xz+y^2+yz+z^2+x+y+z+1.

Constructors

Glex (MonImpl v)

type GlexPoly k v = Vect k (Glex v)

A type representing polynomials with Glex term ordering.

glexvar :: v -> GlexPoly Q v

glexvar v creates a variable in the algebra of commutative polynomials over Q with Glex term ordering. It is provided as a shortcut, to avoid having to provide a type annotation, as with var. For example, the following code creates variables called x, y and z:

[x,y,z] = map glexvar ["x","y","z"]

newtype Grevlex v

A type representing monomials with Grevlex ordering.

Grevlex stands for graded reverse lexicographic. Thus monomials are ordered first by degree, then by reverse lexicographic order. For example, in Grevlex ordering, monomials up to degree two would be ordered as follows: x^2+xy+y^2+xz+yz+z^2+x+y+z+1.

In general, Grevlex leads to the smallest Groebner bases.

Constructors

Grevlex (MonImpl v)

type GrevlexPoly k v = Vect k (Grevlex v)

A type representing polynomials with Grevlex term ordering.

grevlexvar :: v -> GrevlexPoly Q v

grevlexvar v creates a variable in the algebra of commutative polynomials over Q with Grevlex term ordering. It is provided as a shortcut, to avoid having to provide a type annotation, as with var. For example, the following code creates variables called x, y and z:

[x,y,z] = map grevlexvar ["x","y","z"]

data Elim2 a b

Constructors

Elim2 !a !b

bind :: (Eq k, Num k, MonomialConstructor m, Ord a, Show a, Algebra k a) => Vect k (m v) -> (v -> Vect k a) -> Vect k a

Given (Num k, MonomialConstructor m), then Vect k (m v) is the free commutative algebra over v. As such, it is a monad (in the mathematical sense). The following pseudo-code (not legal Haskell) shows how this would work:

instance (Num k, Monomial m) => Monad (\v -> Vect k (m v)) where
    return = var
    (>>=) = bind

bind corresponds to variable substitution, so v `bind` f returns the result of making the substitutions encoded in f into v.

Note that the type signature is slightly more general than that required by (>>=). For a monad, we would only require:

bind :: (MonomialConstructor m, Num k, Ord (m v), Show (m v), Algebra k (m v)) =>
    Vect k (m u) -> (u -> Vect k (m v)) -> Vect k (m v)

Instead, the given type signature allows us to substitute in elements of any algebra. This is occasionally useful.

bind performs variable substitution

flipbind :: (MonomialConstructor m, Algebra k b, Show b, Eq k, Ord b, Num k) => (t -> Vect k b) -> Vect k (m t) -> Vect k b

eval :: (Eq k, Num k, MonomialConstructor m, Eq (m v), Show v) => Vect k (m v) -> [(Vect k (m v), k)] -> k

Evaluate a polynomial at a point. For example eval (x^2+y^2) [(x,1),(y,2)] evaluates x^2+y^2 at the point (x,y)=(1,2).

subst :: (Eq k, Num k, MonomialConstructor m, Eq (m u), Show u, Ord (m v), Show (m v), Algebra k (m v)) => Vect k (m u) -> [(Vect k (m u), Vect k (m v))] -> Vect k (m v)

Perform variable substitution on a polynomial. For example subst (x*z-y^2) [(x,u^2),(y,u*v),(z,v^2)] performs the substitution x -> u^2, y -> u*v, z -> v^2.

vars :: (Num k, Ord k, MonomialConstructor m, Ord (m v)) => Vect k (m v) -> [Vect k (m v)]

List the variables used in a polynomial

lt :: Vect t1 t -> (t, t1)

lm :: Vect b c -> c

lc :: Vect c a -> c

deg :: Monomial m => Vect t1 m -> Int

toMonic :: (Fractional k, Algebra k b, Show b, Ord b, Eq k) => Vect k b -> Vect k b

tdivides :: Monomial m => (m, t1) -> (m, t) -> Bool

tdiv :: (Fractional t, Monomial t1) => (t1, t) -> (t1, t) -> (t1, t)

tgcd :: (Num t, Monomial t3) => (t3, t2) -> (t3, t1) -> (t3, t)

tmult :: (Num t, Mon t1) => (t1, t) -> (t1, t) -> (t1, t)

(*->) :: (Mon b, Num k) => (b, k) -> Vect k b -> Vect k b infixl 7

quotRemMP :: (Fractional t, Monomial m, Algebra t m, Ord m, Eq t) => Vect t m -> [Vect t m] -> ([Vect t m], Vect t m)

rewrite :: (Fractional t, Monomial m, Algebra t m, Ord m, Eq t) => Vect t m -> [Vect t m] -> Vect t m

(%%) :: (Eq k, Fractional k, Monomial m, Ord m, Algebra k m) => Vect k m -> [Vect k m] -> Vect k m infixl 7

f %% gs is the reduction of a polynomial f with respect to a list of polynomials gs. In the case where the gs are a Groebner basis for an ideal I, then f %% gs is the equivalence class representative of f in R/I, and is zero if and only if f is in I.

Orphan instances

(Eq k, Fractional k, Monomial m, Ord m, Algebra k m) => Fractional (Vect k m)
As a convenience, a partial instance of Fractional is defined for polynomials. The instance is well-defined only for scalars, and gives an error if used on other values. The purpose of this is to allow entry of fractional scalars, in expressions such as x/2. On the other hand, an expression such as 2/x will return an error.

Methods

(/) :: Vect k m -> Vect k m -> Vect k m

recip :: Vect k m -> Vect k m

fromRational :: Rational -> Vect k m