Math.Core.Utils

Description

A module of simple utility functions which are used throughout the rest of the library

Documentation

toSet :: Ord a => [a] -> [a]

sortDesc :: Ord a => [a] -> [a]

insertDesc :: Ord a => a -> [a] -> [a]

setUnionAsc :: Ord a => [a] -> [a] -> [a]

The set union of two ascending lists. If both inputs are strictly increasing, then the output is their union and is strictly increasing. The code does not check that the lists are strictly increasing.

setUnionDesc :: Ord a => [a] -> [a] -> [a]

intersectAsc :: Ord a => [a] -> [a] -> [a]

The (multi-)set intersection of two ascending lists. If both inputs are strictly increasing, then the output is the set intersection and is strictly increasing. If both inputs are weakly increasing, then the output is the multiset intersection (with multiplicity), and is weakly increasing.

multisetSumAsc :: Ord a => [a] -> [a] -> [a]

The multiset sum of two ascending lists. If xs and ys are ascending, then multisetSumAsc xs ys == sort (xs++ys). The code does not check that the lists are ascending.

multisetSumDesc :: Ord a => [a] -> [a] -> [a]

The multiset sum of two descending lists. If xs and ys are descending, then multisetSumDesc xs ys == sortDesc (xs++ys). The code does not check that the lists are descending.

diffAsc :: Ord a => [a] -> [a] -> [a]

The multiset or set difference between two ascending lists. If xs and ys are ascending, then diffAsc xs ys == xs \ ys, and diffAsc is more efficient. If xs and ys are sets (that is, have no repetitions), then diffAsc xs ys is the set difference. The code does not check that the lists are ascending.

diffDesc :: Ord a => [a] -> [a] -> [a]

The multiset or set difference between two descending lists. If xs and ys are descending, then diffDesc xs ys == xs \ ys, and diffDesc is more efficient. If xs and ys are sets (that is, have no repetitions), then diffDesc xs ys is the set difference. The code does not check that the lists are descending.

isSubsetAsc :: Ord a => [a] -> [a] -> Bool

isSubMultisetAsc :: Ord a => [a] -> [a] -> Bool

elemAsc :: Ord a => a -> [a] -> Bool

Is the element in the ascending list?

With infinite lists, this can fail to terminate. For example, elemAsc 1 [12,34,7/8..] would fail to terminate. However, with a list of Integer, this will always terminate.

notElemAsc :: Ord a => a -> [a] -> Bool

Is the element not in the ascending list? (With infinite lists, this can fail to terminate.)

picks :: [a] -> [(a, [a])]

Return all the ways to "pick one and leave the others" from a list

pairs :: [t] -> [(t, t)]

ordpair :: Ord t => t -> t -> (t, t)

foldcmpl :: (b -> b -> Bool) -> [b] -> Bool

isWeaklyIncreasing :: Ord t => [t] -> Bool

isStrictlyIncreasing :: Ord t => [t] -> Bool

isWeaklyDecreasing :: Ord t => [t] -> Bool

isStrictlyDecreasing :: Ord t => [t] -> Bool

cmpfst :: Ord a => (a, b1) -> (a, b) -> Ordering

eqfst :: Eq a => (a, b1) -> (a, b) -> Bool

fromBase :: (Num a, Foldable t) => a -> t a -> a

powersetdfs :: [a] -> [[a]]

Given a set xs, represented as an ordered list, powersetdfs xs returns the list of all subsets of xs, in lex order

powersetbfs :: [a] -> [[a]]

Given a set xs, represented as an ordered list, powersetbfs xs returns the list of all subsets of xs, in shortlex order

combinationsOf :: Int -> [a] -> [[a]]

Given a positive integer k, and a set xs, represented as a list, combinationsOf k xs returns all k-element subsets of xs. The result will be in lex order, relative to the order of the xs.

choose :: Integral a => a -> a -> a

choose n k is the number of ways of choosing k distinct elements from an n-set

class FinSet x where

The class of finite sets

Minimal complete definition

elts

Methods

elts :: [x]

class HasInverses a where

A class representing algebraic structures having an inverse operation. Note that in some cases not every element has an inverse.

Minimal complete definition

inverse

Methods

inverse :: a -> a

(^-) :: (Num a, HasInverses a, Integral b) => a -> b -> a infix 8

A trick: x^-1 returns the inverse of x