Math.NumberTheory.Prime

Description

A module providing functions to test for primality, and find next and previous primes.

Documentation

primes :: [Integer]

A (lazy) list of the primes

isTrialDivisionPrime :: Integer -> Bool

isMillerRabinPrime :: (Integral a, Random a) => a -> Bool

isPrime :: Integer -> Bool

Is this number prime? The algorithm consists of using trial division to test for very small factors, followed if necessary by the Miller-Rabin probabilistic test.

notPrime :: Integer -> Bool

prevPrime :: Integer -> Integer

Given n, prevPrime n returns the greatest p, p < n, such that p is prime

nextPrime :: Integer -> Integer

Given n, nextPrime n returns the least p, p > n, such that p is prime