Math.Projects.ChevalleyGroup.Classical

Documentation

numPtsAG :: (Num a, Integral b) => b -> a -> a

numPtsPG :: (Integral b, Integral a) => b -> a -> a

sl :: FiniteField k => Int -> [k] -> [[[k]]]

The special linear group SL(n,Fq), generated by elementary transvections, returned as matrices

elemTransvection :: (Num t, Num t1, Eq t1, Enum t1) => t1 -> (t1, t1) -> t -> [[t]]

l :: (FiniteField k, Ord k) => Int -> [k] -> [Permutation [k]]

The projective special linear group PSL(n,Fq) == A(n,Fq) == SL(n,Fq)/Z, returned as permutations of the points of PG(n-1,Fq). This is a finite simple group provided n>2 or q>3.

orderL :: Integral a => a -> a -> a

sp2 :: FiniteField k => Int -> [k] -> [[[k]]]

The symplectic group Sp(2n,Fq), returned as matrices

s2 :: (FiniteField k, Ord k) => Int -> [k] -> [Permutation [k]]

The projective symplectic group PSp(2n,Fq) == Cn(Fq) == Sp(2n,Fq)/Z, returned as permutations of the points of PG(2n-1,Fq). This is a finite simple group for n>1, except for PSp(4,F2).

s :: (Ord k, FiniteField k) => Int -> [k] -> [Permutation [k]]

orderS2 :: (Integral b, Integral a) => b -> a -> a

orderS :: (Integral a, Integral b) => b -> a -> a

omegaeven :: FiniteField a => Int -> t -> [[[a]]]

d :: (Ord a, FiniteField a) => Int -> [a] -> [Permutation [a]]

omegaodd :: (FiniteField a1, Foldable t) => Int -> t a -> [[[a1]]]

b :: (Ord a, FiniteField a) => Int -> [a] -> [Permutation [a]]

o :: (FiniteField a, Ord a) => Int -> [a] -> [Permutation [a]]